MATE-131 : [TEST] goods 패키지 분리에 따른 테스트 코드 분리
